safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicyviolation-block-cross-origin-image-expected.txt

CONSOLE MESSAGE: Refused to load the image 'http://localhost:8000/security/resources/abe.png' because it violates the following Content Security Policy directive: "img-src 'none'".

Check that a SecurityPolicyViolationEvent strips detail from cross-origin blocked URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Kicking off the tests:
PASS window.e.documentURI is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/securitypolicyviolation-block-cross-origin-image.html"
PASS window.e.referrer is ""
PASS window.e.blockedURI is "http://localhost:8000"
PASS window.e.violatedDirective is "img-src 'none'"
PASS window.e.effectiveDirective is "img-src"
PASS window.e.originalPolicy is "img-src 'none'"
PASS window.e.sourceFile is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/securitypolicyviolation-block-cross-origin-image.html"
PASS window.e.lineNumber is 23
PASS successfullyParsed is true

TEST COMPLETE
